Additional change from rev 12388 (maxlag not integer) to fix compiler error
git-svn-id: https://svn.code.sf.net/p/autowikibrowser/code/AWB@12391 5227b50d-9349-4574-b055-4b0a7525e11c

+++ AWB/Main.cs
@@ -572,7 +572,7 @@ namespace AutoWikiBrowser
             PageLoaded(page);
         }
 
-        private void MaxlagExceeded(AsyncApiEdit sender, int maxlag, int retryAfter)
+        private void MaxlagExceeded(AsyncApiEdit sender, double maxlag, int retryAfter)
         {
             Retries++;
